+++ src/app/components/channel-table/channel-table.component.spec.ts
@@ -10,7 +10,6 @@ import { EMPTY } from 'rxjs';
 import { of } from 'rxjs/internal/observable/of';
 import { delay, startWith } from 'rxjs/operators';
 // @ts-ignore
-import * as Web3 from 'web3';
 import { CdkDetailRowDirective } from '../../directives/cdk-detail-row.directive';
 import { Channel } from '../../models/channel';
 import { UserToken } from '../../models/usertoken';
@@ -30,51 +29,11 @@ import { TokenInputComponent } from '../token-input/token-input.component';
 import { TokenNetworkSelectorComponent } from '../token-network-selector/token-network-selector.component';
 
 import { ChannelTableComponent } from './channel-table.component';
-import { BatchManager } from '../../services/batch-manager';
+import { StatusPipe } from '../../pipes/status.pipe';
+import { AddressBookService } from '../../services/address-book.service';
+import { stub } from '../../../testing/stub';
+import { MockConfig } from '../../../testing/mock-config';
 import Spy = jasmine.Spy;
-
-class MockWeb3 extends Web3 {
-    isChecksum = false;
-    checksumAddress = '';
-
-    constructor() {
-        super();
-
-        const mockWeb3 = this;
-        this.eth.getBlockNumber = function () {
-            return Promise.resolve(3694423);
-        };
-        this.utils.checkAddressChecksum = function () {
-            return mockWeb3.isChecksum;
-        };
-
-        this.utils.toChecksumAddress = function () {
-            return mockWeb3.checksumAddress;
-        };
-    }
-
-}
-
-export class MockConfig extends RaidenConfig {
-    public web3: Web3 = new MockWeb3();
-    private testBatchManager: BatchManager = new BatchManager(this.web3.currentProvider);
-
-    get batchManager(): BatchManager {
-        return this.testBatchManager;
-    }
-
-    private get mock(): MockWeb3 {
-        return (this.web3 as MockWeb3);
-    }
-
-    updateChecksumAddress(address: string): void {
-        this.mock.checksumAddress = address;
-    }
-
-    setIsChecksum(isChecksum: boolean): void {
-        this.mock.isChecksum = isChecksum;
-    }
-}
 
 describe('ChannelTableComponent', () => {
     let component: ChannelTableComponent;
@@ -83,7 +42,12 @@ describe('ChannelTableComponent', () => {
     let refreshingSpy: Spy;
     let tokenSpy: Spy;
 
+    let addressBookStub: AddressBookService;
+
     beforeEach(async(() => {
+        addressBookStub = stub<AddressBookService>();
+        addressBookStub.get = () => ({});
+
         TestBed.configureTestingModule({
             declarations: [
                 ChannelTableComponent,
@@ -95,6 +59,7 @@ describe('ChannelTableComponent', () => {
                 KeysPipe,
                 SubsetPipe,
                 DecimalPipe,
+                StatusPipe,
                 CdkDetailRowDirective,
                 TokenNetworkSelectorComponent
             ],
@@ -103,6 +68,10 @@ describe('ChannelTableComponent', () => {
                 {
                     provide: RaidenConfig,
                     useClass: MockConfig
+                },
+                {
+                    provide: AddressBookService,
+                    useFactory: () => addressBookStub
                 },
                 RaidenService,
                 ChannelPollingService,
