+++ src/app/components/token-network-selector/token-network-selector.component.spec.ts
@@ -12,7 +12,7 @@ import { TokenPipe } from '../../pipes/token.pipe';
 import { RaidenConfig } from '../../services/raiden.config';
 import { RaidenService } from '../../services/raiden.service';
 import { SharedService } from '../../services/shared.service';
-import { MockConfig } from '../channel-table/channel-table.component.spec';
+import { MockConfig } from '../../../testing/mock-config';
 
 import { TokenNetworkSelectorComponent } from './token-network-selector.component';
 import { errorMessage, mockEvent } from '../../../testing/interaction-helper';
