+++ src/app/services/raiden.service.spec.ts
@@ -2,7 +2,7 @@ import { HttpClientModule } from '@angular/common/http';
 import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
 import { fakeAsync, flush, inject, TestBed, tick } from '@angular/core/testing';
 import { of } from 'rxjs';
-import { MockConfig } from '../components/channel-table/channel-table.component.spec';
+import { MockConfig } from '../../testing/mock-config';
 import { Channel } from '../models/channel';
 import { UserToken } from '../models/usertoken';
 import { RaidenConfig } from './raiden.config';
