+++ src/testing/stub.ts
@@ -1,0 +1,10 @@
+export function stub<T>(): T {
+    const typeAssertion = <T>{};
+    for (const prop in typeAssertion) {
+        if (typeAssertion.hasOwnProperty(prop)) {
+            typeAssertion[prop] = undefined;
+        }
+    }
+
+    return typeAssertion;
+}
